Update the logic to avoid passing the interface
Unfortunately based on the release the out of classrun is based on a different interface. Dynamic calls seem to lead to issues in the console so as a simple workaround I moved the output to the main method where you can directly use out. All entries are written to an internal string table before. As the console output is anyhow only at the end this should not have any relevant effect.